Move retile and activateView functions to better places
