+++ src/main/kotlin/gh/marad/tiler/tiler/internal/WindowsTiler.kt
@@ -1,19 +1,22 @@
 package gh.marad.tiler.tiler.internal
 
+import gh.marad.tiler.common.*
+import gh.marad.tiler.common.Window
+import gh.marad.tiler.common.filteringrules.FilteringRules
 import gh.marad.tiler.tiler.internal.views.ViewManager
 import gh.marad.tiler.tiler.internal.views.ViewSwitcher
-import gh.marad.tiler.common.DesktopState
-import gh.marad.tiler.common.MinimizeWindow
-import gh.marad.tiler.common.TilerCommand
-import gh.marad.tiler.common.Window
+import gh.marad.tiler.common.layout.LayoutSpace
+import gh.marad.tiler.os.OsFacade
 import gh.marad.tiler.tiler.TilerFacade
+import gh.marad.tiler.tiler.internal.views.View
 import gh.marad.tiler.common.Window as TilerWindow
 
 class WindowsTiler(
     private val viewManager: ViewManager,
-    private val getDesktopState: () -> DesktopState
+    private val filteringRules: FilteringRules,
+    private val os: OsFacade,
 ): TilerFacade {
-    private val viewSwitcher = ViewSwitcher(viewManager, getDesktopState)
+    private val viewSwitcher = ViewSwitcher(viewManager, this::getDesktopState)
     override var enabled = true
 
     override fun initializeWithOpenWindows(): List<TilerCommand> {
@@ -29,12 +32,12 @@ class WindowsTiler(
 
     override fun switchToView(viewId: Int): List<TilerCommand> {
         if (!enabled) return emptyList()
-        return viewSwitcher.switchToView(viewId)
+        return viewSwitcher.switchToView(viewId) + retile()
     }
 
     override fun switchToPreviousView(): List<TilerCommand> {
         if (!enabled) return emptyList()
-        return viewSwitcher.switchToPreviousView()
+        return viewSwitcher.switchToPreviousView() + retile()
     }
 
     override fun addWindow(window: Window): List<TilerCommand> {
@@ -61,7 +64,30 @@ class WindowsTiler(
     override fun retile(): List<TilerCommand> {
         if (!enabled) return emptyList()
         val view = viewManager.currentView()
-        return retile(view, getDesktopState().windowsToManage, getDesktopState().layoutSpace)
+        val desktopState = getDesktopState()
+        return retile(view, desktopState.windowsToManage, desktopState.layoutSpace)
     }
 
+    private fun getDesktopState(): DesktopState = os.getDesktopState(filteringRules)
+
+    private fun retile(view: View, allWindows: Windows, space: LayoutSpace): List<TilerCommand> {
+        allWindows.filterNot { it.isMinimized }
+            .forEach { view.addWindow(it.id) }
+        val windowsInView = view.filterWindowsInView(allWindows)
+        view.updateWindowsInView(windowsInView.map { it.id })
+        val positionedWindows = view.layout.retile(windowsInView, space)
+        return createPositionCommands(allWindows, positionedWindows)
+    }
+
+    private fun createPositionCommands(allWindows: Windows, positionedWindows: Windows): List<TilerCommand> {
+        val positionedWindowsById = positionedWindows.associateBy { it.id }
+        val commands = mutableListOf<TilerCommand>()
+        allWindows.forEach {
+            val positionedWindow = positionedWindowsById[it.id]
+            if (positionedWindow != null && it.position != positionedWindow.position) {
+                commands.add(SetWindowPosition(it.id, positionedWindow.position))
+            }
+        }
+        return commands
+    }
 }
