+++ src/test/kotlin/gr/marad/tiler/core/CoreTest.kt
@@ -5,7 +5,6 @@ import gh.marad.tiler.common.layout.LayoutSpace
 import gh.marad.tiler.tiler.internal.views.View
 import gh.marad.tiler.tiler.internal.views.ViewManager
 import gh.marad.tiler.common.*
-import gh.marad.tiler.tiler.internal.createPositionCommands
 import io.kotest.matchers.equals.shouldNotBeEqual
 import io.kotest.matchers.shouldBe
 import io.kotest.property.Arb
@@ -47,21 +46,21 @@ val windowGen = arbitrary { rs ->
     Window(windowIdGen.next(rs), windowTitles.next(rs), classNames.next(rs), "exe_path", posGen.next(rs), isMinimized = false, isMaximized = false, isPopup = false)
 }
 
-class CoreTest {
-    @Test
-    fun `should set window position if it changes`() {
-        // given
-        val aWindow = windowGen.next()
-        val updatedPosition = WindowPosition(x = 0, y = 0, width = 200, height = 100)
-
-        // when
-        val (command) = createPositionCommands(listOf(aWindow), listOf(aWindow.copy(position = updatedPosition)))
-
-        // then
-        command shouldBe SetWindowPosition(aWindow.id, updatedPosition)
-
-    }
-}
+//class CoreTest {
+//    @Test
+//    fun `should set window position if it changes`() {
+//        // given
+//        val aWindow = windowGen.next()
+//        val updatedPosition = WindowPosition(x = 0, y = 0, width = 200, height = 100)
+//
+//        // when
+//        val (command) = createPositionCommands(listOf(aWindow), listOf(aWindow.copy(position = updatedPosition)))
+//
+//        // then
+//        command shouldBe SetWindowPosition(aWindow.id, updatedPosition)
+//
+//    }
+//}
 
 
 class ViewTest {
